DocumentsIntoVectors/requirements.txt
langchain
python-dotenv
openai
pypdf
sentence transformers
faiss-gpu
python-telegram-bot